Cleanup pydantic-ai implementation
