+++ src/fastomop/main.py
@@ -1,44 +1,56 @@
 """Main entry point for FastOMOP application."""
 
+from doctest import FAIL_FAST
 import uvicorn
 import asyncio
 from fastomop import __version__
 from fastomop.otel import tracer
 from opentelemetry.trace import SpanKind
+from fastomop.agents.supervisor import FastOmopSupervisor
+from rich.console import Console
+from rich.table import Table
+
+console = Console()
 
 
-def run_a2a():
-    from fastomop.agents.sql_agent import app as sqlagent_app
-
-    uvicorn.run(sqlagent_app, port=8000)
-
-
-async def run_graph():
-    # This function is for running the orchestrator graph
-    from fastomop.agents.supervisor import create_graph
-
-    graph = await create_graph()
-
-    async def stream_graph_updates(user_input: str):
-        async for event in graph.astream(
-            {"messages": [{"role": "user", "content": user_input}]}
-        ):
-            for value in event.values():
-                print("Assistant:", value["messages"])
-
-    while True:
-        user_input = input("User: ")
-        if user_input.lower() in ["quit", "exit", "q"]:
-            print("Goodbye!")
-            break
-        await stream_graph_updates(user_input)
 
 
 @tracer.start_as_current_span(
     name=__name__, kind=SpanKind.SERVER, attributes={"version": __version__}
 )
 async def main_async() -> None:
-    await run_graph()
+    """Entry point for the fastomop command."""
+    supervisor = FastOmopSupervisor()
+    print("FastOMOP v2 - pydantic-AI implementation")
+    print("----------------------------------------")
+    print("Type 'quit', 'exit' or 'q' to quit")
+    print("----------------------------------------")
+
+    while True:
+        try:
+            user_query = input("User: ")
+            if user_query.lower() in ["quit", "exit", "q"]:
+                print("Goodbye!")
+                break
+
+            if not user_query:
+                continue
+
+            print ("Processing query...")
+            result = await supervisor.process_query(user_query)
+
+            if result.success:
+                print(f"\n{result.get_summary()}")
+                print(f"\nAssistant: {result.final_answer}")
+            else:
+                print(f"\nError: {result.final_answer}")
+
+        except KeyboardInterrupt:
+            print("\nGoodbye!")
+            break
+        except Exception as e:
+            print(f"\nError: {e}")
+
 
 
 def main() -> None:
